Wrong usage of schema for array in "highlights"
The type property itself is not a string, but the current specification claims that. Actually, it is an `object` with `properties`.

You can see that this must be changed just one level above. In the `work` array, it has been done correctly.

I think this is nothing to discuss. Just wrong semantics.